Unify cascade detection logic
