+++ crypto_screener/config/gu_config.py
@@ -43,6 +43,10 @@ class GuConfig:
     CASCADE_PULLBACK_NEXT_RATIO_MAX: float = 0.9
     # Минимальная глубина первых откатов в долях ATR.
     CASCADE_MIN_PULLBACK_NATR: float = 2.5
+    # Минимальная длительность отката в барах.
+    CASCADE_MIN_PULLBACK_BARS: int = 3
+    # Минимальный разрыв между касаниями в барах.
+    CASCADE_MIN_GAP_BARS: int = 5
     # Минимальное соотношение отступа поддержки от низа проторговки к размеру проторговки.
     SUPPORT_CONSOLIDATION_RATIO_MIN: float = 0.3
     # Минимальная длительность окна группировки каскадов.
